requirements.txt
# API y HTTP
requests
amadeus

# Variables de entorno (configuración sensible)
python-dotenv

# Base de datos PostgreSQL
psycopg2-binary
sqlalchemy

# Utilidades para logging y validación
python-dateutil
pytz

# Scheduling de tareas periódicas (ejecutar cada X horas)
schedule

# Análisis de datos y cálculos estadísticos
numpy
pandas
# Notificaciones por Telegram
python-telegram-bot
